Rename backend-wrapper to os-integration
