+++ app/frontend/Shared/ContextMenu.ts
@@ -3,7 +3,7 @@ import { blobToDataURL, NotImplemented, type URLString } from "../../logic/util/
 import { gt } from "../../l10n/l10n";
 import { appGlobal } from "../../logic/app";
 import { sanitize } from "../../../lib/util/sanitizeDatatypes";
-import { openExternalURL } from "../../logic/util/backend-wrapper";
+import { openExternalURL } from "../../logic/util/os-integration";
 
 /**
  * Handles the Electron `<webview>` `"context-menu"` event.
